Add configuration files that specify which Ruby version to use and which gemset to use.
These configuration files are useful because they will cause an
automatic switch to the correct environment when the project folder is
entered. This avoids a lot of subtle bugs introduced by being in the
incorrect environment.

.ruby-version specifies the version of Ruby that should be used in the
project.

.ruby-gemset specifies which gemset to use, both for installing new
gemset and as a source of gemset locations.

Both .ruby-gemset and .ruby-version are compatible with other Ruby
Version Managers such as rbenv and chruby.