Add setup to our test suite.
We have added a few setup items to our test suite. In particular, we've
created a simple app that we can instantiate our middleware with.

We've created an instance of our Middleware with our simple app, so that
it is ready for testing.

We've created a method called stubbed_quotes that will reach into the
Quotes module and pull out the array of quotes defined there.

We've created a request method that uses an instance of the
Rack::MockRequest object instantiated with our middleware to avoid
making actual HTTP calls.

Finally, note that we are using allow_any_instance_of to stub out the
call to QuoteMiddleware#quotes. Normally, we'd use
allow(QuoteMiddleware) instead, but this does not work for private
methods such as QuoteMiddleware#quotes.